requirements.txt
streamlit>=1.30
soprano-tts>=0.1.0
